feat(IAM): add Role and User repositories with essential queries
- Implemented `IRoleRepository` with `ExistsByName` method.
- Added `IUserRepository` with methods for user retrieval, password verification, and role/workshop filtering.
- Created `RoleRepository` to check if a role exists by name.
- Developed `UserRepository` with `FindByUsernameAsync`, `ExistsByUsername`, `VerifyPassword`, and `FindAllByRoleAndWorkshopAsync`.

+++ YARA.WorkshopNGine.API/Shared/Domain/Repositories/IBaseRepository.cs
@@ -3,7 +3,7 @@ namespace YARA.WorkshopNGine.API.Shared.Domain.Repositories;
 public interface IBaseRepository<TEntity>
 {
     Task AddAsync(TEntity entity);
-    Task<TEntity?> FindByIdAsync(int id);
+    Task<TEntity?> FindByIdAsync(long id);
     void Update(TEntity entity);
     void Remove(TEntity entity);
     Task<IEnumerable<TEntity>> ListAsync();
